--- Tepok Nyamuk/GameViewController.swift
@@ -1,50 +1,0 @@
-//
-//  GameViewController.swift
-//  Tepok Nyamuk
-//
-//  Created by Benaya Oktavianus on 09/06/20.
-//  Copyright © 2020 -. All rights reserved.
-//
-
-import UIKit
-import SpriteKit
-import GameplayKit
-
-class GameViewController: UIViewController {
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        if let view = self.view as! SKView? {
-            // Load the SKScene from 'GameScene.sks'
-            if let scene = SKScene(fileNamed: "MainMenuScene") {
-                // Set the scale mode to scale to fit the window
-                scene.scaleMode = .aspectFill
-                
-                // Present the scene
-                view.presentScene(scene)
-            }
-            
-            view.ignoresSiblingOrder = true
-//            view.showsPhysics = true
-            view.showsFPS = true
-            view.showsNodeCount = true
-        }
-    }
-    
-    override var shouldAutorotate: Bool {
-        return false
-    }
-    
-    override var supportedInterfaceOrientations: UIInterfaceOrientationMask {
-        if UIDevice.current.userInterfaceIdiom == .phone {
-            return .allButUpsideDown
-        } else {
-            return .all
-        }
-    }
-    
-    override var prefersStatusBarHidden: Bool {
-        return true
-    }
-}
